requirements.txt
alembic==1.4.3
attrs==20.3.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
embeddings==0.0.8
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.1.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
tqdm==4.53.0
tweepy==3.9.0
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
